Main refactoring
