+++ lib/view/graphs_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Graphs extends StatelessWidget {
+  const Graphs({
+    Key? key,
+    required this.context,
+  }) : super(key: key);
+
+  final BuildContext context;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(color: Colors.purple);
+  }
+}
