+++ pubspec.yaml
@@ -29,11 +29,13 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  provider: ^5.0.0
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  
   google_fonts: ^2.2.0
   html: ^0.15.0
   shared_preferences: ^2.0.12
